+++ src/net/ssehub/kernel_haven/metric_haven/metric_components/weights/MultiWeight.java
@@ -32,8 +32,8 @@ public class MultiWeight implements IVariableWeight {
     }
     
     @Override
-    public int getWeight(String variable) {
-        int weight = 1;
+    public long getWeight(String variable) {
+        long weight = 1;
         for (int i = 0; i < weights.length; i++) {
             weight *= weights[i].getWeight(variable);
         }
@@ -42,7 +42,7 @@ public class MultiWeight implements IVariableWeight {
     }
     
     @Override
-    public int getWeight(String variable, File codefile) {
+    public long getWeight(String variable, File codefile) {
         int weight = 1;
         for (int i = 0; i < weights.length; i++) {
             weight *= weights[i].getWeight(variable, codefile);
